Refactored code

+++ OrganisationProfitCalculator/OrganisationProfitCalculator.Data/Models/Office.cs
@@ -4,6 +4,7 @@
     {
         public string Name { get; set; }
         public string Parent { get; set; }
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
     }   
 }
+
